Fix empty result sets
It seems like find_by_sql(...) sometimes returns an empty result set
when it shouldn't.

gems/dm-ar-finders-1.2.0/lib/dm-ar-finders.rb:129 fetches each row from
the database by calling next!() on a reader until the return value is
nil.  It's hypothesised that next!() is returning false due to some
error, instead of due to reaching the end of the results set.

Ruby's next!() is do_mysql_cReader_next() in ext/do_mysql/do_mysql.c.

It apparently returns false early if @reader is nil or it refers to a
null pointer.  It clearly returns false for both normal case (end of
result set) and error.

We change it here to throw exception on error instead of return false.